requirements.txt
beautifulsoup4==4.5.3
conda==4.2.13
cycler==0.10.0
dj-database-url==0.4.1
Django==1.11
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gunicorn==19.6.0
h5py==2.7.1
imageio==2.3.0
Keras==2.1.6
lxml==4.2.1
matplotlib==2.0.0
mock==2.0.0
numpy==1.14.3
pandas==0.23.0
pbr==4.0.3
Pillow==5.1.0
psycopg2==2.7.1
pycosat==0.6.1
pycrypto==2.6.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
requests==2.11.1
ruamel-yaml===-VERSION
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
subprocess32==3.2.7
virtualenv==15.1.0
